refactor: load secrets from env
