+++ .env.example
@@ -8,6 +8,13 @@ SUPABASE_JWT_SECRET=xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
 NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=pk_test_xxxxxxx
 STRIPE_SECRET_KEY=sk_test_xxxxxxx
 STRIPE_WEBHOOK_SECRET=whsec_xxxxxxx
+# Pricing IDs
+STRIPE_BASIC_MONTH_PRICE_ID=price_basic_month
+STRIPE_BASIC_YEAR_PRICE_ID=price_basic_year
+STRIPE_BASIC_PRODUCT_ID=prod_basic
+STRIPE_PRO_MONTH_PRICE_ID=price_pro_month
+STRIPE_PRO_YEAR_PRICE_ID=price_pro_year
+STRIPE_PRO_PRODUCT_ID=prod_pro
 
 #email server
 EMAIL_SERVER_USER=xxxxxx
