+++ config.ts
@@ -1,59 +1,67 @@
-// This file is automatically generated by the system and contains the Stripe configuration for the project
-// Note: In production environments, please use environment variables to store this sensitive information
+/**
+ * Application configuration loaded from environment variables.
+ * Sensitive Stripe identifiers are kept in env to avoid committing secrets.
+ */
 
+function getEnv(name: string): string {
+  const value = process.env[name];
+  if (!value) {
+    throw new Error(`Missing environment variable: ${name}`);
+  }
+  return value;
+}
 
 const config = {
-	metadata: {
-		title: "Saas Starter",
-		description: "Saas Starter Kit for demo purpose ",
-		keywords: ["Saas", "Starter Kit", "Demo", "Note Taking", "AI", "Chatbot"],
-	},
-	theme: {
-		colors: {
-			primary: '#5059FE',
-			primaryHover: '#4048ed',
-			border: '#E5E7EB',  // Standard border color
-			borderHover: '#D1D5DB'  // Border hover color
-		}
-	},
-	// 加入 stripe plan 的   name 和 description
-	stripe: {
-		free: {
-			monthPrice: 0,
-			yearPrice: 0,
-			monthPriceId: '',
-			yearPriceId: '',
-			productId: '',
-			name: 'Free',
-			description: 'Free plan',
-		},
-		basic: {
-			monthPrice: 9.9,
-			yearPrice: 90,
-			monthPriceId: 'price_1RVpwVCk4zpvQXEQuDdS4RWT',
-			yearPriceId: 'price_1RVqNGCk4zpvQXEQaqcQC3RR',
-			productId: 'prod_SQhLl2ALMvkaQ3',
-			name: 'Basic',
-			description: 'Basic plan',
-		},
-		pro: {
-			monthPrice: 29.9,
-			yearPrice: 280,
-			monthPriceId: 'price_1RVqKGCk4zpvQXEQOZrS1Wbs',
-			yearPriceId: 'price_1RVqMMCk4zpvQXEQi0QctdyG',
-			productId: 'prod_SQhj8DablxYrZ0',
-			name: 'Pro',
-			description: 'Pro plan',
-		},
-	},
-	appName: "saas starter kit",
-	socialLinks: {
-		github: "https://github.com",
-		twitter: "https://twitter.com",
-		linkedin: "https://linkedin.com"
-	},
-	emailProvider: "nodemailer",
+  metadata: {
+    title: "Saas Starter",
+    description: "Saas Starter Kit for demo purpose ",
+    keywords: ["Saas", "Starter Kit", "Demo", "Note Taking", "AI", "Chatbot"],
+  },
+  theme: {
+    colors: {
+      primary: '#5059FE',
+      primaryHover: '#4048ed',
+      border: '#E5E7EB',
+      borderHover: '#D1D5DB',
+    },
+  },
+  // Stripe plan configuration
+  stripe: {
+    free: {
+      monthPrice: 0,
+      yearPrice: 0,
+      monthPriceId: '',
+      yearPriceId: '',
+      productId: '',
+      name: 'Free',
+      description: 'Free plan',
+    },
+    basic: {
+      monthPrice: 9.9,
+      yearPrice: 90,
+      monthPriceId: getEnv('STRIPE_BASIC_MONTH_PRICE_ID'),
+      yearPriceId: getEnv('STRIPE_BASIC_YEAR_PRICE_ID'),
+      productId: getEnv('STRIPE_BASIC_PRODUCT_ID'),
+      name: 'Basic',
+      description: 'Basic plan',
+    },
+    pro: {
+      monthPrice: 29.9,
+      yearPrice: 280,
+      monthPriceId: getEnv('STRIPE_PRO_MONTH_PRICE_ID'),
+      yearPriceId: getEnv('STRIPE_PRO_YEAR_PRICE_ID'),
+      productId: getEnv('STRIPE_PRO_PRODUCT_ID'),
+      name: 'Pro',
+      description: 'Pro plan',
+    },
+  },
+  appName: "saas starter kit",
+  socialLinks: {
+    github: "https://github.com",
+    twitter: "https://twitter.com",
+    linkedin: "https://linkedin.com",
+  },
+  emailProvider: "nodemailer",
 };
 
-
-export default config
+export default config;
